data/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
motor==3.7.0
pymongo==4.12.1
pydantic-settings==2.9.1
python-dotenv==1.1.0
email-validator==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
